test/tier1/lakebridge_reconcile/pyproject.toml
[project]
name = "lakebridge-reconcile-tier1"
version = "0.1.0"
description = "Tier 1 skill logic tests for the Lakebridge Reconcile skill"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.0",
    "sqlglot>=28.0",
    "openai",
    "pyyaml>=6.0",
    "pytest>=8.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
markers = [
    "tier1: Skill logic tests (mock tools, requires Databricks Claude Sonnet)",
]
testpaths = ["."]
